requirements/dev.txt
-r base.txt

ipython
django-debug-toolbar
django-extensions
factory_boy
